Client/requirements.txt
cycler==0.10.0
decorator==4.4.2
fingerprint-enhancer==0.0.11
gobject==0.1.0
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.21.0
opencv-python==4.4.0.44
Pillow==8.2.0
protobuf==3.6.0
pycairo==1.20.1
PyGObject==3.40.1
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.2
scipy==1.7.0
six==1.11.0
tifffile==2021.6.14
